Don't render title as list item
This fixes an issue where "12. April" would get rendered as "1. April" because
it looks like a Markdown list item to our renderer. Now, we parse titles as
titles, instead of standalone text, which causes the renderer to give us the
results we want. This also adds some basic tests for the applyBasicMarkdown()
func.

Closes #470